+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
-- Feature: `Label`: Add `pill` props.
+- Feature: `Label`: Add `dot` props.
  
 ## [5.0.2] - 2021-10-13